Update html css & Js
Added the volume control feature and drop here text in html. Added the glow and scale animation for the dropped icons in CSS with some little bit of changing in the shape of dropbox and for JS fixed the bug for the play button which was playing sound of all the music even without dropping the icon in dropbox and also added the volume control feature for each icons.